bug-hunt #565: ext4_extents - drop 32-bit LBA truncation in ext4_ext_write_block
block_num * (block_size/512) can exceed 2^32 sectors on large
filesystems (2TB @ 512B sectors, 16TB @ 4K blocks); casting the
64-bit lba to uint32_t wrapped metadata writes to the wrong sector.
blockdev_write_sectors takes uint64_t, so pass the full value.
Matches the ext4_read_block fix from bug-hunt #564.

+++ src/fs/ext4_extents.c
@@ -323,7 +323,11 @@ static int ext4_ext_write_block(struct ext4_priv *ep, uint32_t block_num,
     int ret;
 
     for (uint32_t i = 0; i < sectors; i++) {
-        if (blockdev_write_sectors(ep->dev_id, (uint32_t)(lba + i), 1,
+        /* lba is 64-bit: block_num * sectors-per-block can exceed 2^32
+         * sectors (2TB on 512-byte-sector disks) on large filesystems.
+         * Truncating it to uint32_t would wrap the write to the wrong
+         * sector.  blk_submit_sync/blockdev_write_sectors take uint64_t. */
+        if (blockdev_write_sectors(ep->dev_id, lba + i, 1,
                                    buf + i * 512) != 0)
             return -EIO;
     }
